# src/types/response/admin.response.ts
export interface saveupdateresadmin{
    _id:string;
    firstname:string;
    lastname:string;
    cell:string;
    designation:string;
    address:string;
    joindate:string;
    createdat:string;
    updatedat:string
}